# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bff_user",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message lisible de l'erreur",
            "example": "Invalid credentials provided."
          },
          "error": {
            "description": "Détail technique optionnel de l'erreur"
          }
        },
        "required": [
          "message"
        ]
      },
      "LoginView": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Adresse email de l'utilisateur",
            "example": "alice.dupont@mairie360.fr"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "description": "Mot de passe de l'utilisateur",
            "example": "MotDePasse123"
          },
          "device_info": {
            "type": "string",
            "description": "Informations sur le périphérique utilisé pour se connecter (chaîne vide acceptée, comme par Core API)",
            "example": "Firefox"
          }
        },
        "required": [
          "email",
          "password",
          "device_info"
        ]
      },
      "RegisterView": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Adresse email du nouvel utilisateur",
            "example": "alice.dupont@mairie360.fr"
          },
          "first_name": {
            "type": "string",
            "minLength": 1,
            "description": "Prénom du nouvel utilisateur",
            "example": "Alice"
          },
          "last_name": {
            "type": "string",
            "minLength": 1,
            "description": "Nom du nouvel utilisateur",
            "example": "Dupont"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "description": "Mot de passe du nouvel utilisateur",
            "example": "MotDePasse123"
          },
          "phone_number": {
            "type": [
              "string",
              "null"
            ],
            "description": "Numéro de téléphone du nouvel utilisateur",
            "example": "+33123456789"
          }
        },
        "required": [
          "email",
          "first_name",
          "last_name",
          "password"
        ]
      },
      "ForceChangePasswordView": {
        "type": "object",
        "properties": {
          "new_password": {
            "type": "string",
            "minLength": 1,
            "description": "Nouveau mot de passe de l'utilisateur",
            "example": "NouveauMotDePasse123"
          },
          "token": {
            "type": "string",
            "minLength": 1,
            "description": "Token de changement forcé du mot de passe",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "new_password",
          "token"
        ]
      },
      "AboutResponseView": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Adresse email de l'utilisateur",
            "example": "alice.dupont@mairie360.fr"
          },
          "first_name": {
            "type": "string",
            "description": "Prénom de l'utilisateur",
            "example": "Alice"
          },
          "last_name": {
            "type": "string",
            "description": "Nom de l'utilisateur",
            "example": "Dupont"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ],
            "description": "Numéro de téléphone de l'utilisateur (null si non renseigné)",
            "example": "+33123456789"
          },
          "status": {
            "type": "string",
            "description": "Statut du compte utilisateur",
            "example": "active"
          }
        },
        "required": [
          "email",
          "first_name",
          "last_name",
          "phone",
          "status"
        ]
      },
      "AuthTokenResponse": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "Refresh token retourné par le Core API",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refresh_token"
        ]
      },
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Logged out successfully"
          }
        },
        "required": [
          "message"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Authentifie un utilisateur",
        "description": "Transmet les identifiants au Core API et retourne le token JWT.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Utilisateur authentifié ; le JWT d’accès est dans Authorization, le corps contient le refresh token.",
            "headers": {
              "Authorization": {
                "description": "Bearer <access token>",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Identifiants invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "Première connexion : mot de passe à changer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "token"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Core API indisponible ou réponse amont invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Crée un utilisateur",
        "description": "Valide puis transmet les informations d'inscription au Core API (POST /api/v1/auth/register, route publique).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterView"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé avec succès"
          },
          "400": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Utilisateur déjà existant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Core API indisponible ou réponse amont invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/force_change_password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Force le changement de mot de passe",
        "description": "Transmet le token et le nouveau mot de passe au Core API sur /api/v1/auth/force_change_password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForceChangePasswordView"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Mot de passe changé avec succès"
          },
          "400": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token invalide ou expiré",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Token de première connexion inconnu ou expiré",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Core API indisponible ou réponse amont invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Déconnecte un utilisateur",
        "description": "Supprime le cookie HTTP-only contenant le token d'accès.",
        "responses": {
          "200": {
            "description": "Utilisateur déconnecté",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}/about": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Récupère les informations publiques d'un utilisateur",
        "description": "Transmet la session (en-tête Authorization, x-session-token ou cookie accessToken) au Core API sur /api/v1/user/{id}/ et ne renvoie que les informations publiques.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Identifiant numérique de l'utilisateur",
              "example": 42
            },
            "required": true,
            "description": "Identifiant numérique de l'utilisateur",
            "name": "userId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Informations utilisateur récupérées avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AboutResponseView"
                }
              }
            }
          },
          "400": {
            "description": "Identifiant utilisateur invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Utilisateur non authentifié ou ID invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Core API indisponible ou réponse amont invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}